Resurrect TestQgsPostgresDomain
This test wasn't being added via ADD_PYTHON_TEST, so wasn't being
run
