+++ tests/src/python/test_provider_postgres.py
@@ -1138,5 +1138,27 @@ class TestPyQgsPostgresProviderCompoundKey(unittest.TestCase, ProviderTestCase):
         return set([])
 
 
+class TestQgsPostgresDomain(unittest.TestCase):
+
+    @classmethod
+    def setUpClass(cls):
+        """
+        Setup the involved layer
+        :return:
+        """
+        cls.dbconn = 'service=\'qgis_test\''
+        if 'QGIS_PGTEST_DB' in os.environ:
+            cls.dbconn = os.environ['QGIS_PGTEST_DB']
+        # Create test layer
+        cls.vl = QgsVectorLayer(cls.dbconn + ' sslmode=disable key=\'pk\' table="qgis_test"."colors" sql=', 'colors', 'postgres')
+
+        QgsProject.instance().addMapLayer(cls.vl)
+
+    def test_postgres_domain(self):
+        self.assertEqual(self.vl.dataProvider().enumValues(1), ['red', 'green', 'blue'])
+        self.assertEqual(self.vl.dataProvider().enumValues(2), ['yellow', 'cyan', 'magenta'])
+        self.assertEqual(self.vl.dataProvider().enumValues(3), ['Alchemilla', 'Alstroemeria', 'Alyssum'])
+
+
 if __name__ == '__main__':
     unittest.main()
